+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrGraphGenerator.kt
@@ -300,7 +300,7 @@ internal class IrGraphGenerator(
                             it
                           }
                         }
-                        .let(::IrTypeKey)
+                        .let(IrTypeKey.Companion::invoke)
                     irExprBody(
                       irInvoke(
                         dispatchReceiver =