+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrTypeKey.kt
@@ -8,14 +8,12 @@ import dev.zacsweers.metro.compiler.graph.BaseTypeKey
 import dev.zacsweers.metro.compiler.unsafeLazy
 import org.jetbrains.kotlin.ir.types.IrSimpleType
 import org.jetbrains.kotlin.ir.types.IrType
-import org.jetbrains.kotlin.ir.types.isMarkedNullable
 import org.jetbrains.kotlin.ir.types.typeOrFail
-import org.jetbrains.kotlin.ir.types.typeOrNull
-import org.jetbrains.kotlin.ir.util.render
 
 // TODO cache these in DependencyGraphTransformer or shared transformer data
 @Poko
-internal class IrTypeKey(override val type: IrType, override val qualifier: IrAnnotation? = null) :
+internal class IrTypeKey
+private constructor(override val type: IrType, override val qualifier: IrAnnotation?) :
   BaseTypeKey<IrType, IrAnnotation, IrTypeKey> {
 
   private val cachedRender by unsafeLazy { render(short = false, includeQualifier = true) }
@@ -38,27 +36,13 @@ internal class IrTypeKey(override val type: IrType, override val qualifier: IrAn
         append(" ")
       }
     }
-    val typeString =
-      if (short) {
-        type.renderShort()
-      } else {
-        type.render()
-      }
-    append(typeString)
+    type.renderTo(this, short)
   }
 
-  private fun IrType.renderShort(): String = buildString {
-    append(simpleName)
-    if (isMarkedNullable()) {
-      append("?")
-    }
-    if (this@renderShort is IrSimpleType) {
-      arguments
-        .takeUnless { it.isEmpty() }
-        ?.joinToString(", ", prefix = "<", postfix = ">") {
-          it.typeOrNull?.renderShort() ?: "<error>"
-        }
-        ?.let { append(it) }
+  companion object {
+    operator fun invoke(type: IrType, qualifier: IrAnnotation? = null): IrTypeKey {
+      // Canonicalize on the way through
+      return IrTypeKey(type.canonicalize(), qualifier)
     }
   }
 }